requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
weaviate-client==4.5.0
PyPDF2==3.0.1
langchain==0.1.9
python-magic==0.4.27
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.34
celery==5.3.6
redis==5.0.1 